har gjort så man ikke kan pille ved andres ønsker

+++ src/main/java/apiassignment/wishlist/controller/WishlistController.java
@@ -60,15 +60,10 @@ public class WishlistController {
             return "login";
         }
         User user = (User) session.getAttribute("user");
-        if(!wishlistService.isLoggedIn(session)) {
-            return "redirect:/login";
-        }
         List<Wishlist> wishLists = wishlistService.getAllWishlistsByUserId(user.getUserId());
 
         //Capitalize first character in name
         String name = user.getName().substring(0, 1).toUpperCase() + user.getName().substring(1);
-
-
 
         model.addAttribute("wishlists", wishLists);
         model.addAttribute("name", name);
@@ -77,7 +72,17 @@ public class WishlistController {
 
     @GetMapping("/wishlist/{id}")
     public String getWishlist (@PathVariable("id") int wishlistId, Model model, HttpSession session) {
+        if(!wishlistService.isLoggedIn(session)) {
+            return "login";
+        }
         User user = (User) session.getAttribute("user");
+        if(!wishlistService.checkWishlistIdWithUserId(wishlistId, user.getUserId())){
+            return "/profile";
+        }
+        if(wishlistService.checkWishlistIdWithUserIdWithoutFriends(wishlistId, user.getUserId())){
+            model.addAttribute("myWishes", true);
+
+        }
         Wishlist wishlist = wishlistService.getWishlistById(wishlistId);
         List<Wish> wishes = wishlistService.getAllWishesFromWishlistId(wishlistId);
 
@@ -114,7 +119,6 @@ public class WishlistController {
 
     @GetMapping("/create/wishlist")
     public String createWishlist(Model model, HttpSession session) {
-
         if(!wishlistService.isLoggedIn(session)) {
             return "login";
         }
@@ -128,6 +132,9 @@ public class WishlistController {
 
     @PostMapping("/create/wishlist")
     public String saveWishList(@ModelAttribute Wishlist wishlist, HttpSession session) {
+        if(!wishlistService.isLoggedIn(session)) {
+            return "login";
+        }
         User user = (User) session.getAttribute("user");
         if(user == null) {
             return "redirect:/login";
@@ -138,7 +145,15 @@ public class WishlistController {
     }
 
     @PostMapping("/wishlist/delete/{id}")
-    public String deleteWishlist(@PathVariable int id) {
+    public String deleteWishlist(@PathVariable int id, HttpSession session) {
+        if(!wishlistService.isLoggedIn(session)) {
+            return "login";
+        }
+        User user = (User) session.getAttribute("user");
+        //check if it's my profile or an admin
+        if(!wishlistService.checkWishlistIdWithUserId(id, user.getUserId())){
+            return "login";
+        }
         wishlistService.deleteAllWishesWithWishlistId(id);
         wishlistService.deleteWishlist(id);
         return "redirect:/profile";
@@ -167,19 +182,25 @@ public class WishlistController {
 
     @PostMapping("/delete/{id}")
     public String deleteUser(@PathVariable int id, HttpSession session) {
-        User user = (User) session.getAttribute("user");
-
         if(!wishlistService.isLoggedIn(session)) {
             return "redirect:/login";
         }
-
-        wishlistService.deleteUser(id);
+        User user = (User) session.getAttribute("user");
+        if(user.getUserId() == id || user.getRoleId() == 1) { //tjekker om det er ens egen bruger eller man er admin
+            wishlistService.deleteUser(id);
+        }
         return "redirect:/login";
     }
 
     @GetMapping("/wish/add/{wishlistId}")
     public String addWish(@PathVariable int wishlistId, Model model, HttpSession session) {
+        if(!wishlistService.isLoggedIn(session)) {
+            return "redirect:/login";
+        }
         User user = (User) session.getAttribute("user");
+         if(!wishlistService.checkWishlistIdWithUserId(wishlistId, user.getUserId())){
+            return "profile";
+        }
         if (user == null) {
             return "redirect:/login";
         }
@@ -191,6 +212,10 @@ public class WishlistController {
 
     @PostMapping("wish/save")
     public String saveWish(@ModelAttribute Wish wish, HttpSession session) {
+        if(!wishlistService.isLoggedIn(session)) {
+            return "redirect:/login";
+
+        }
         User user = (User) session.getAttribute("user");
         if(user == null) {
             return "redirect:/login";
@@ -201,9 +226,15 @@ public class WishlistController {
 
     @GetMapping("wish/{id}/edit")
     public String editWish(Model model, @PathVariable int id, HttpSession session) {
+        if(!wishlistService.isLoggedIn(session)) {
+            return "redirect:/login";
+        }
         User user = (User) session.getAttribute("user");
         if (user == null) {
             return "redirect:/login";
+        }
+        if(!wishlistService.checkWishIdWithUserId(id, user.getUserId())){
+            return "redirect:/profile";
         }
         Wish wish = wishlistService.getWishById(id);
         if (wish == null) {
@@ -215,9 +246,15 @@ public class WishlistController {
 
     @PostMapping("/wish/update")
     public String updateWish(@ModelAttribute Wish wish, HttpSession session) {
+        if(!wishlistService.isLoggedIn(session)) {
+            return "redirect:/login";
+        }
         User user = (User) session.getAttribute("user");
         if (user == null) {
             return "redirect:/login";
+        }
+        if(!wishlistService.checkWishIdWithUserId(wish.getWishId(), user.getUserId())){
+            return "redirect:/profile";
         }
         wishlistService.updateWish(wish);
         return "redirect:/wishlist/" + wish.getWishlistId();
@@ -225,9 +262,15 @@ public class WishlistController {
 
     @PostMapping("/wish/delete/{id}")
     public String deleteWish(@PathVariable int id, HttpSession session) {
+        if(!wishlistService.isLoggedIn(session)) {
+            return "redirect:/login";
+        }
+        User user = (User) session.getAttribute("user");
+        if(!wishlistService.checkWishIdWithUserId(id, user.getUserId())){
+            return "redirect:/profile";
+        }
         Wish wish = wishlistService.getWishById(id);
         wishlistService.deleteWish(id);
-
         return "redirect:/wishlist/" + wish.getWishlistId();
     }
 
@@ -236,12 +279,19 @@ public class WishlistController {
         if(!wishlistService.isLoggedIn(session)) {
             return "redirect:/login";
         }
+        User user = (User) session.getAttribute("user");
+        if(!wishlistService.checkWishIdWithUserId(id, user.getUserId())){ //Checker om sessionId matcher med wishId
+            return "redirect:/profile";
+        }
+        if(wishlistService.checkWishIdWithUserIdWithoutFriends(id, user.getUserId())){
+            model.addAttribute("myWishes", true);
+
+        }
         Wish wish = wishlistService.getWishById(id);
         if(wish == null) {
             return "redirect:/profile";
         }
         model.addAttribute("wish", wish);
-
 
 
         return "wish";
@@ -358,7 +408,7 @@ public class WishlistController {
 
 
     @PostMapping("/add/friend/{username}")
-    public String addFriend(@PathVariable String username, Model model, HttpSession session){
+    public String addFriend(@PathVariable String username, HttpSession session){
         if(!wishlistService.isLoggedIn(session)) {
             return "login";
         }
@@ -425,6 +475,7 @@ public class WishlistController {
 
     @GetMapping("/share/{token}")
     public String shareWishlist(@PathVariable String token, Model model, HttpSession session) {
+
         User loggedUser = (User) session.getAttribute("user");
         boolean notLoggedIn;
 
@@ -457,14 +508,11 @@ public class WishlistController {
         if(loggedUser == null) {
             notLoggedIn = true;
         }
-
         model.addAttribute("token", token);
         model.addAttribute("notLoggedIn",notLoggedIn);
-
         Wish wish = wishlistService.getWishById(id);
         model.addAttribute("wish", wish);
         model.addAttribute("isReserved", wishlistService.isWishReservedById(id));
-
         return "sharedWish";
     }
 
@@ -472,14 +520,10 @@ public class WishlistController {
 
     @PostMapping("/shared/wish/{token}/{id}")
     public String updateIsReserved (@PathVariable int id, @PathVariable String token, HttpSession session) {
-        System.out.println("Updating wish with ID: " + id);
-
         if (wishlistService.isLoggedIn(session)) {
             wishlistService.updateIsReservedByWishId(id);
-            System.out.println("Wish updated successfully.");
             return "redirect:/shared/wish/" + token +"/" + id;
         } else {
-            System.out.println("User not logged in, redirecting to login.");
             return "redirect:/login";
         }
 
@@ -488,13 +532,14 @@ public class WishlistController {
 
     @GetMapping("/admin/update/{id}")
     public String adminUpdateUser(@PathVariable int id, HttpSession session, Model model) {
-        User loggedUser = (User) session.getAttribute("user");
-        User getUser = wishlistService.getUserById(id);
         if(!wishlistService.isLoggedIn(session)) {
             return "redirect:/login";
-        } else if (loggedUser.getRoleId() != 1) {
+        }
+        User loggedUser = (User) session.getAttribute("user");
+        if (loggedUser.getRoleId() != 1) {
             return "redirect:/login";
         }
+        User getUser = wishlistService.getUserById(id);
         model.addAttribute("user", getUser);
         model.addAttribute("updatedUser", new User());
         return "adminUpdateUser";
@@ -502,13 +547,13 @@ public class WishlistController {
 
     @PostMapping("/admin/update")
     public String adminUpdateUser(@ModelAttribute User updatedUser, HttpSession session){
-        User loggedUser = (User) session.getAttribute("user");
         if(!wishlistService.isLoggedIn(session)) {
             return "redirect:/login";
-        } else if (loggedUser.getRoleId() != 1) {
+        }
+        User loggedUser = (User) session.getAttribute("user");
+        if (loggedUser.getRoleId() != 1) {
             return "redirect:/login";
         }
-
         wishlistService.updateUser(updatedUser);
 
         return "redirect:/admin";
